% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_clust.R
\name{em_clust}
\alias{em_clust}
\title{Expectation-Maximization Algorithm}
\usage{
em_clust(data_set, k)
}
\arguments{
\item{data_set}{The data set to perform the Expectation-Maximization algorithm. All columns of data set must be quantitative.}

\item{k}{Will choose k number of random observations in the data set as the starting points of the means of your respective normal distributions}
}
\value{
A list the parameters means, variances, the probabilities of cluster membership, and the final clusters each observation is assigned to
}
\description{
Implements the Expectation-Maximization algorithm.
}
